+++ src/app/_components/utils/NetworkErrorNotification.tsx
@@ -3,7 +3,7 @@
 import showToast from '@/utils/showToast';
 import { useEffect } from 'react';
 
-export default function NetworkErrorPage() {
+export default function NetworkErrorNotification() {
   useEffect(() => {
     showToast('네트워크 연결이 끊겼습니다.', 'error');
   }, []);
